+++ structures/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'structures.apps.StructuresConfig'
 
-__version__ = '1.2.0a4'
+__version__ = '1.2.0a5'
 __title__ = 'Alliance Structures'