+++ CHANGELOG.md
@@ -9,13 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.2.0] - 2020-04-12
 
-**POS fuel & sovereignty info**
+**POS fuel & system type tags**
 
 If you are upgrading from a version prior to 1.1.0 please make sure to follow the upgrade instructions for 1.1.0.
 
 ### Added
 
 - Shows "fuel expires" for starbases ([#20](https://gitlab.com/ErikKalkoken/aa-structures/issues/20))
+- Shows a system type tag, e.g. "highsec" or "nullsec"
 - Shows sov tag if a system of structure owner has sov for that system
 
 ## [1.1.4] - 2020-04-14